refactor: reorganize server utility imports and add MCP server configuration functions
