+++ apps/electron/src/renderer/components/mcp/server/Manual.tsx
@@ -23,7 +23,7 @@ import {
 import {
   validateMcpServerJson,
   processMcpServerConfigs,
-} from "@/main/modules/mcp-core/server/shared/mcp-server-utils";
+} from "./utils/mcp-server-utils";
 import { toast } from "sonner";
 import { Textarea } from "@mcp_router/ui";
 import { Alert, AlertDescription, AlertTitle } from "@mcp_router/ui";
